dcrpg: meta table versioning
update meta best block and ibd_complete flag

Add bestBlockHash to metaData struct and meta table.

return int64 height from DeleteBestBlock and DeleteBestBlocks, never
sql.ErrNoRows.

Add RetrieveBestBlock that returns int64 height (-1 when no rows), and
never sql.ErrNoRows.

remove table comments after upgrade from legacy versioning

Check best block referenced in transactions table too.

docs and cleanup
