+++ db/dcrpg/pgblockchain_charts_test.go
@@ -46,7 +46,7 @@ func openDB() (func() error, error) {
 	}
 	var err error
 	db, err = NewChainDB(&dbi, &chaincfg.MainNetParams, nil, true, true, addrCacheCap,
-		nil, new(parserInstance))
+		nil, new(parserInstance), nil)
 	cleanUp := func() error { return nil }
 	if db != nil {
 		cleanUp = db.Close
